refactor: move context package
