+++ src/main/java/com/github/floington500/api/controller/commands/UpdateFile.java
@@ -1,7 +1,7 @@
 package com.github.floington500.api.controller.commands;
 
 import com.github.floington500.common.command.FileOperation;
-import com.github.floington500.common.command.context.FileContext;
+import com.github.floington500.common.context.FileContext;
 import com.github.floington500.common.exceptions.FileUploadException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
