+++ src/main/java/com/github/floington500/api/controller/commands/UploadFile.java
@@ -1,12 +1,11 @@
 package com.github.floington500.api.controller.commands;
 
 import com.github.floington500.common.command.FileOperation;
-import com.github.floington500.common.command.context.FileContext;
+import com.github.floington500.common.context.FileContext;
 import com.github.floington500.common.exceptions.FileUploadException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
