+++ src/main/java/com/github/floington500/api/controller/download/FileDownloadController.java
@@ -1,6 +1,6 @@
 package com.github.floington500.api.controller.download;
 
-import com.github.floington500.common.command.context.FileContextFactory;
+import com.github.floington500.common.context.FileContextFactory;
 import com.github.floington500.common.command.handler.FileHandlerImpl;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.AllArgsConstructor;
