+++ src/main/java/com/github/floington500/common/command/FileOperation.java
@@ -1,9 +1,8 @@
 package com.github.floington500.common.command;
 
-import com.github.floington500.common.command.context.FileContext;
+import com.github.floington500.common.context.FileContext;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
