+++ src/main/java/com/github/floington500/common/command/Operation.java
@@ -1,6 +1,6 @@
 package com.github.floington500.common.command;
 
-import com.github.floington500.common.command.context.FileContext;
+import com.github.floington500.common.context.FileContext;
 import com.github.floington500.common.command.handler.FileHandler;
 import org.springframework.http.ResponseEntity;
