+++ src/main/java/com/github/floington500/common/command/handler/FileHandler.java
@@ -1,7 +1,7 @@
 package com.github.floington500.common.command.handler;
 
 import com.github.floington500.common.command.Operation;
-import com.github.floington500.common.command.context.FileContext;
+import com.github.floington500.common.context.FileContext;
 import org.springframework.http.ResponseEntity;
 
 import java.util.Optional;
